% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestCutOff.R
\name{findBestCuttoff}
\alias{findBestCuttoff}
\title{genLinearPred function}
\usage{
findBestCuttoff(Data, time = "time", ind = "ind", PRINT = TRUE)
}
\arguments{
\item{Data}{matrix}

\item{betas}{named list of beta values}
}
\description{
genLinearPred function
}
\examples{
## Pass the function a matrix and a named list of beta values
## The beta values must have the same names as the rows in the matrix
genLinearPred(betas, matrix)
}
\keyword{linear}
\keyword{predictor}
