% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLinearPred.R
\name{genLinearPred}
\alias{genLinearPred}
\title{genLinearPred function}
\usage{
genLinearPred(Data)
}
\arguments{
\item{Data}{object of class LinearPredictor}
}
\description{
genLinearPred function
}
\examples{
## Pass the function a matrix and a named list of beta values
## The beta values must have the same names as the rows in the matrix
genLinearPred(Data)

## Returns the class object with the linear predictor added to the ExpressionSet slot
## Also filters out all genes that are not in either the beta values of the ExpressionSet
}
\keyword{linear}
\keyword{predictor}
