% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classLinPred.R
\docType{class}
\name{LinearPredictor-class}
\alias{LinearPredictor-class}
\alias{LinearPredictor}
\title{This class will store the data being tested.}
\arguments{
\item{ExprSet}{ExpressionSet}

\item{betas}{named list of beta values.}

\item{time_col}{character object giving the column name for the follow up time}

\item{ind_col}{character object giving the column name for the event indicator}

\item{Plots}{list of plots}
}
\description{
This class will store the data being tested.
}
\examples{
x <- LinearPredictor()

}
\keyword{Linear}
\keyword{Predictors}
\keyword{classification}
