% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLinearPred.R
\name{genLinearPred}
\alias{genLinearPred}
\title{genLinearPred function}
\usage{
genLinearPred(Data)
}
\arguments{
\item{Data}{matrix}

\item{betas}{named list of beta values}
}
\description{
genLinearPred function
}
\examples{
## Pass the function a matrix and a named list of beta values
## The beta values must have the same names as the rows in the matrix
genLinearPred(betas, matrix)
}
\keyword{linear}
\keyword{predictor}
